Completed Building Lists and Navigation
Completed tutorial 2 of 3 for Chapter 1: SwiftUI Essentials.

Now the Landmarks app decodes `landmarkData.json` into an Array<Landmark> -- no Landmark data is hard-coded. Additional assets (pictures) of landmarks have been added.

The app begins from a LandmarkList -- which contains a NavigationView of NavigationLink. Each NavigationLink consists of a LandmarkRow -- which has a thumbnail/title of the landmark.

Upon selecting a NavigationLink, we end up on a LandmarkDetail page, which is was the first part of the project, prior to this commit.

I've decided to remove the ScrollView from the text section -- now the entire body wrapped in a ScrollView.